Initial mailing list
The intent is to contact a cross-section of interested parties.

Each envelope contained two double-sided pages where the first page included a table of contents and the mailing list. On the outside of the envelope was printed the table of contents in the return address area and on the back was printed the mailing list.